Fix a compile error when using a constant to define an array's size (#369)
* Fix a compile error when using a constant to define an array's size.

When using `#[derive(DdsType)]` on a type like `struct ArrayContainingType([u8; ARRAY_LENGTH]);` led to a compile error in `get_type_identifier`. The problem is that the member `array_bound_seq` of  `PlainArraySElemDefn` and `PlainArrayLElemDefn` is of type `u8` or `u32`, respectively. When using an actual integer to define the arrays size, the integer expression is used in the proc macro and the compiler happily casts (?) it to the right integer type. But when using a `usize` constant, the constant expression is put into the `vec![...]` call and the compiler fails, as it expects a smaller integer type.

This patch adds proper casts to handle this situation as well as a compile time test case.

* Turn checked casts into normal `as` casts.